requirements.txt
shodan
PyYAML
requests
colorama
